Zmienione typy

+++ Projekt pk4/space invaders ++/player.cpp
@@ -3,11 +3,11 @@
 
 
 
-Player::Player():Player(Vector2f(640,360),100)
+Player::Player():Player(Vector2i(640,360),100)
 {
 }
 
-Player::Player(Vector2f setposition, int sethp)
+Player::Player(Vector2i setposition, int sethp)
 {
 	position = setposition;
 	hp = sethp;
@@ -18,14 +18,14 @@ Player::~Player()
 {
 }
 
-void Player::SetPosition(Vector2f newposition)
+void Player::SetPosition(Vector2i newposition)
 {
 	position = newposition;
 }
 
-void Player::SetPosition(Vector2i newposition)
+void Player::SetPosition(Vector2f newposition)
 {
-	position =(Vector2f)newposition;
+	position =(Vector2i)newposition;
 }
 
 void Player::SetHP(int newhp)
@@ -33,7 +33,7 @@ void Player::SetHP(int newhp)
 	hp = newhp;
 }
 
-sf::Vector2f Player::GetPosition()
+sf::Vector2i Player::GetPosition()
 {
 	return position;
 }
